2d_attractor
Plot of the attractor for the two-dimensional map f(x) = (ax − y^2, by + xy). a and b are constants, where 0 < a < 1 and b > 1. Defines the function 'attractor' which sets the initial values for x and y, skips through the first 200 iterations and appends the next N points to the list 'plot' so that the points plotted converge close to the attractor.